Frontend: + added game service for axios requests + added components Cards & Lists for Games, SavedGames, & Reviews + updated package.json for MUI integration
