--- frontend/src/components/games/GameCard.tsx
@@ -1,17 +1,0 @@
-import {Game} from "../../models/Game";
-import {MouseEventHandler} from "react";
-
-export default function GameCard({game}: {game: Game}) {
-
-    const onCardPressed: MouseEventHandler<HTMLButtonElement> = (e) => {
-        e.preventDefault()
-
-        console.log(`Open game page for ${game.name}`)
-    }
-
-    return (
-        <button onClick={onCardPressed}>
-            {game.id} - {game.name}
-        </button>
-    )
-}
