+++ frontend/src/games/SavedGamesList.tsx
@@ -1,9 +1,10 @@
-import {SavedGame} from "../../models/SavedGame";
+import {SavedGame} from "./SavedGame";
 import SavedGameCard from "./SavedGameCard";
+import "../styles/Styles.css";
 
 export default function SavedGamesList({games}: {games: SavedGame[]}) {
     return (
-        <ul>
+        <ul className={"SavedGamesList"}>
             {games.map(game => (<li key={game.id}><SavedGameCard savedGame={game}/></li>))}
         </ul>
     )
